requirements.txt
streamlit
typing
textblob
groq
python-dotenv